terraform/modules/lambda/src/requirements.txt
boto3
tweepy

openai
pyyaml
pydantic
beautifulsoup4